requirements.txt
PyPDF2
sentence-transformers
python-dotenv
azure-search-documents
streamlit
openai
tenacity
langchain
tiktoken
streamlit_chat
azure-storage-blob
PyMuPDF